Refactored useForm hook to useFormPersist hook to save form values in localstorage
